Migrate fjs/effects/node/module.f.ts to .f.mjs, split type-level API into types.ts
The Node.js effect operations module: filesystem, networking, exec,
console, sandbox, test registration, and the NodeOp/NodeProgram types.
Nearly the entire public surface moves to types.ts unchanged (real
TypeScript); runtime combinators become JSDoc-typed .mjs, using
@template/@param on writeFromStream and both where an inline @type
can't expose its generics to the body, and an any-cast on both's
all(a, b) call to carry over the original's unchecked `as` widening.

Updates 26 dependents across the repo (djs, text/sgr, cli, dev, ci,
media/type, cas, mcp, protocol/mcp, emergent_testing, website, and
effects' own submodules) to import runtime values from .f.mjs and
types from types.ts separately, plus a batch of stale module.f.ts doc
references (including three with stale line-number citations that now
split across module.f.mjs and types.ts). proof.f.ts stays TypeScript
for now.
