+++ fjs/cas/proof.f.ts
@@ -5,7 +5,8 @@ import { computeSync, sha256 } from '../crypto/sha2/module.f.mjs'
 import { fileCas, casAddFile, collectRead, type FileCasOperation, casUpload } from './module.f.ts'
 import { match, pure, runPure, step } from '../effects/module.f.mjs'
 import type { Effect } from '../effects/types.ts'
-import { mkdir, writeFile, rm, readFile, type ReadFile, type WriteFile, type Rm, type Mkdir, type IoResult, access } from '../effects/node/module.f.ts'
+import { mkdir, writeFile, rm, readFile, access } from '../effects/node/module.f.mjs'
+import type { ReadFile, WriteFile, Rm, Mkdir, IoResult } from '../effects/node/types.ts'
 import type { Ok } from '../types/result/types.ts'
 import { error, ok } from '../types/result/module.f.mjs'
 import { emptyState, virtual } from '../effects/node/virtual/module.f.ts'
