+++ fjs/ci/module.f.ts
@@ -5,7 +5,8 @@
  */
 import { mapStep, step } from '../effects/module.f.mjs'
 import type { Effect } from '../effects/types.ts'
-import { access, writeUtf8File, type NodeOp } from '../effects/node/module.f.ts'
+import { access, writeUtf8File } from '../effects/node/module.f.mjs'
+import type { NodeOp } from '../effects/node/types.ts'
 import { functionalscript, images } from './config/module.f.mjs'
 import {
     architecture,
