+++ fjs/ci/nix/proof.f.ts
@@ -5,7 +5,7 @@
  */
 import { assert, assertEq } from '../../asserts/module.f.mjs'
 import { step } from '../../effects/module.f.mjs'
-import { readUtf8File } from '../../effects/node/module.f.ts'
+import { readUtf8File } from '../../effects/node/module.f.mjs'
 import { emptyState, virtual } from '../../effects/node/virtual/module.f.ts'
 import { nixpkgs } from '../config/module.f.mjs'
 import { nodeNixJobs } from '../node/module.f.ts'
