+++ fjs/cli/module.f.ts
@@ -1,4 +1,5 @@
-import { errorExit, log, type NodeOp, type NodeProgramOptions, type Write } from '../effects/node/module.f.ts'
+import { errorExit, log } from '../effects/node/module.f.mjs'
+import type { NodeOp, NodeProgramOptions, Write } from '../effects/node/types.ts'
 import { pure, step } from '../effects/module.f.mjs'
 import type { Effect } from '../effects/types.ts'
 import { at, fromEntries } from '../types/object/module.f.mjs'
