+++ fjs/cli/proof.f.ts
@@ -1,5 +1,5 @@
 import { pure } from '../effects/module.f.mjs'
-import type { NodeOp, NodeProgramOptions } from '../effects/node/module.f.ts'
+import type { NodeOp, NodeProgramOptions } from '../effects/node/types.ts'
 import { defaultNodeProgramOptions, emptyState, virtual } from '../effects/node/virtual/module.f.ts'
 import { dispatch, type Commands } from './module.f.ts'
 import { assert, assertEq } from '../asserts/module.f.mjs'
