+++ fjs/djs/transpiler/module.f.ts
@@ -17,7 +17,8 @@ import { type ParseError, parseFromTokens } from '../parser/module.f.ts'
 import { run, type AstModule } from '../ast/module.f.ts'
 import { foldStep, pure, step } from '../../effects/module.f.mjs'
 import type { Effect } from '../../effects/types.ts'
-import { readUtf8File, type ReadFile } from '../../effects/node/module.f.ts'
+import { readUtf8File } from '../../effects/node/module.f.mjs'
+import type { ReadFile } from '../../effects/node/types.ts'
 
 /**
  * State threaded through the recursive transpilation of a DJS module graph.
