+++ fjs/effects/node/todo/readjsonfile-writejsonfile-helpers.md
@@ -41,7 +41,7 @@ export const writeJsonFile = (path: string) => (value: Unknown): Effect<WriteFil
 
 ### Related
 
-- `fjs/effects/node/module.f.ts` — `readFile`/`writeFile` effects.
+- `fjs/effects/node/module.f.mjs` — `readFile`/`writeFile` effects.
 - `fjs/text/module.f.mjs` — `utf8`/`utf8ToString`.
 - [`fjs/media/json/module.f.ts`](../../../media/json/module.f.ts) — `parse` and
   `stringify`, the FunctionalScript pipeline both helpers sit on.
