+++ fjs/emergent_testing/module.f.ts
@@ -11,24 +11,21 @@
  * @module
  */
 import { reset, fgGreen, fgRed, bold, csiWrite } from '../text/sgr/module.f.ts'
-import {
-    all,
-    awaitIfPromise,
-    sandbox,
-    test,
-    type All,
-    type Await,
-    type Env,
-    type NodeProgram,
-    type NodeProgramOptions,
-    type Program,
-    type Sandbox,
-    type SandboxResult,
-    type Test,
-    type TestContext,
-    type Write,
-    type WriteConsoles
-} from '../effects/node/module.f.ts'
+import { all, awaitIfPromise, sandbox, test } from '../effects/node/module.f.mjs'
+import type {
+    All,
+    Await,
+    Env,
+    NodeProgram,
+    NodeProgramOptions,
+    Program,
+    Sandbox,
+    SandboxResult,
+    Test,
+    TestContext,
+    Write,
+    WriteConsoles
+} from '../effects/node/types.ts'
 import { history, historyStep, pure, step } from '../effects/module.f.mjs'
 import type { Effect, Operation } from '../effects/types.ts'
 import { eff } from '../effects/eff/module.f.mjs'
