+++ fjs/media/type/module.f.ts
@@ -39,7 +39,7 @@ import type { Nullable } from '../../types/nullable/types.ts'
 import { pure, step } from '../../effects/module.f.mjs'
 import type { Effect, Operation } from '../../effects/types.ts'
 import type { List } from '../../effects/list/types.ts'
-import type { IoResult } from '../../effects/node/module.f.ts'
+import type { IoResult } from '../../effects/node/types.ts'
 import { ok, error } from '../../types/result/module.f.mjs'
 import { isValidCodePoint, isTextCodePoint } from '../../text/code_point/module.f.mjs'
 import type { Utf8State } from '../../text/utf8/types.ts'
