+++ fjs/module.f.ts
@@ -7,7 +7,8 @@ import { compile } from './djs/module.f.ts'
 import { main as testMain } from './emergent_testing/module.f.ts'
 import { commands as casCommands } from './cas/cli/module.f.ts'
 import { main as ciMain } from './ci/module.f.ts'
-import { import_, type NodeOp, type NodeProgram } from './effects/node/module.f.ts'
+import { import_ } from './effects/node/module.f.mjs'
+import type { NodeOp, NodeProgram } from './effects/node/types.ts'
 import { dispatch, type Commands } from './cli/module.f.ts'
 import { casMcpServer } from './mcp/module.f.ts'
 import { pure, step } from './effects/module.f.mjs'
