+++ fjs/proof.f.ts
@@ -1,6 +1,6 @@
 import { assert, assertEq } from './asserts/module.f.mjs'
 import { pure } from './effects/module.f.mjs'
-import type { NodeProgram, NodeProgramOptions } from './effects/node/module.f.ts'
+import type { NodeProgram, NodeProgramOptions } from './effects/node/types.ts'
 import { defaultNodeProgramOptions, emptyState, virtual, type Dir } from './effects/node/virtual/module.f.ts'
 import { main } from './module.f.ts'
