+++ fjs/protocol/mcp/stdio/module.f.ts
@@ -31,7 +31,8 @@
  */
 import { pure, step } from '../../../effects/module.f.mjs'
 import type { Effect, Operation } from '../../../effects/types.ts'
-import { readLine, write, type IoResult, type Read, type Write } from '../../../effects/node/module.f.ts'
+import { readLine, write } from '../../../effects/node/module.f.mjs'
+import type { IoResult, Read, Write } from '../../../effects/node/types.ts'
 import { tryUtf8 } from '../../../text/module.f.mjs'
 import { parse, stringify, type Unknown } from '../../../media/json/module.f.ts'
 import { sort } from '../../../types/object/module.f.mjs'
